// frontend/utils/types.ts
export type TProduct = {
    id: number;
    created_at: string;
    name: string;
    description: string;
    sku: string;
    sale_price: number;
    low_stock_threshold: number;
    cost_price: number;
    main_image: string;
    category: string;
    effective_price: string;
    slug: string;
    brand: string | null;
    stock_quantity: number;
    is_on_sale: boolean;
    is_new: boolean;
    store_name: string;
    store_slug: string;
    status: string;
    discount?: number | null;
    discount_start?: string | null;
    discount_end?: string | null;
    image_two?: string;
};
export type TProducts = {
    category: string;
    products: {
        product_id: number;
        name: string;
        description: string;
        sale_price: number;
        main_image: string;
        slug: string;
        effective_price: string;
        store_name: string;
        is_on_sale: boolean;
        stock_quantity?: number;
        is_new: boolean;
        discount?: number | null;
        discount_start?: string | null;
        discount_end?: string | null;
    }[];
};
export type TSingleStoreProducts = {
    created_at: string;
    cost_price: number;
    store_slug: string;
    status: string;
    id: number;
    name: string;
    description: string;
    category: string;
    brand: string;
    sale_price: number;
    main_image: string;
    slug: string;
    sku: string;
    effective_price: string;
    store_name: string;
    is_on_sale: boolean;
    stock_quantity: number;
    low_stock_threshold: number;
    is_new: boolean;
    discount?: number | null;
    discount_start?: string | null;
    discount_end?: string | null;
};

export type TCreateProduct = {
    name: string;
    description: string;
    cost_price: number;
    sale_price: number;
    stock_quantity: string | number;
    category: string;
    store_slug: string;
    low_stock_threshold: string | number;
    main_image?: string;
    image_two?: null | string;
};
export type TUpdateProduct = {
    name: string;
    description: string;
    cost_price: number;
    sale_price: number;
    stock_quantity: number;
    category: string;
    slug: string;
    low_stock_threshold: number;
    main_image?: string;
    image_two?: null | string;
};
export type TStores = {
    id: number;
    name: string;
    description: string;
    email: string;
    slug: string;
    phone: string;
    logo: string;
    address_line1: string;
    address_line2: string | null;
    city: string;
    province: string;
    postal_code: string;
};
export type TStoresByName = {
    id: number;
    name: string;
    slug: string;
    description: string;
    email: string;
    phone: string;
    city: string;
    province: string;
    postal_code: string;
    address_line1: string;
    average_rating: number;
    total_reviews: number;
    banner_url?: string;
    meta?: {
        currentPage?: number;
        totalPages?: number;
        totalCount?: number;
    };
    products: {
        id: number;
        name: string;
        description: string;
        category: string;
        sale_price: number;
        main_image: string;
        slug: string;

        effective_price: string;
        store_name: string;
        is_on_sale: boolean;
        stock_quantity?: number;
        is_new: boolean;
        discount?: number | null;
        discount_start?: string | null;
        discount_end?: string | null;
    }[];
};
export type TMyStores = {
    id: number;
    name: string;
    description: string;
    slug: string;
    banner_url: string | null;
};